refactor(migrate): move utilities out of migrate script
